# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API REST para la gestión de la clínica veterinaria: autenticación, usuarios, clientes, empleados, mascotas, citas, pagos y notificaciones.",
        "title": "API Clínica Veterinaria",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Equipo de Soporte API",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "2.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v2",
    "paths": {
...
        "/auth/2fa/disable": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Disables 2FA for the authenticated user. Requires Bearer auth.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Disable two-factor authentication",
                "responses": {
                    "200": {
                        "description": "2FA disabled",
                        "schema": {
                            "$ref": "#/definitions/clinic-vet-api_internal_shared_http.APIResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/clinic-vet-api_internal_shared_http.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/clinic-vet-api_internal_shared_http.APIResponse"
                        }
                    }
                }
            }
        },
        "/auth/2fa/enable": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Enables 2FA for the authenticated user (method: totp, sms, or email). Requires Bearer auth.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Enable two-factor authentication",
                "parameters": [
                    {
                        "description": "2FA method",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/clinic-vet-api_internal_infrastructure_http_handlers_dtos.EnableTwoFactorRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "2FA enabled",
                        "schema": {
                            "$ref": "#/definitions/clinic-vet-api_internal_shared_http.APIResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/clinic-vet-api_internal_shared_http.APIResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/clinic-vet-api_internal_shared_http.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/clinic-vet-api_internal_shared_http.APIResponse"
                        }
                    }
                }
            }
        },
        "/auth/2fa/verify": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Completes login by verifying the two-factor code. Returns new session payload. Requires Bearer auth.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Verify 2FA code",
                "parameters": [
                    {
                        "description": "2FA code",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/clinic-vet-api_internal_infrastructure_http_handlers_dtos.VerifyTwoFactorRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Session with tokens",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/clinic-vet-api_internal_shared_http.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/clinic-vet-api_internal_infrastructure_http_handlers_dtos.SessionResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid code",
                        "schema": {
                            "$ref": "#/definitions/clinic-vet-api_internal_shared_http.APIResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/clinic-vet-api_internal_shared_http.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/clinic-vet-api_internal_shared_http.APIResponse"
                        }
                    }
                }
            }
        },
        "/auth/activate": {
            "post": {
                "description": "Activates a user account using the user ID and activation code (e.g. from email link). Supports query params or JSON body.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Activate account",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "user_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Activation code",
                        "name": "code",
                        "in": "query",
                        "required": true
                    },
                    {
                        "description": "Alternative: JSON body with user_id and code",
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/clinic-vet-api_internal_infrastructure_http_handlers_dtos.ActivateAccountRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Account activated",
                        "schema": {
                            "$ref": "#/definitions/clinic-vet-api_internal_shared_http.APIResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid or missing parameters",
                        "schema": {
                            "$ref": "#/definitions/clinic-vet-api_internal_shared_http.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/clinic-vet-api_internal_shared_http.APIResponse"
                        }
                    }
                }
            }
        },
        "/auth/login": {
            "post": {
                "description": "Authenticates with email and password. Returns access and refresh tokens. If 2FA is enabled, data contains dtos.RequiresTwoFactorResponse instead of the session.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Login",
                "parameters": [
                    {
                        "description": "Email and password; optional two_factor_code",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/clinic-vet-api_internal_infrastructure_http_handlers_dtos.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Login successful (session payload with tokens and user)",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/clinic-vet-api_internal_shared_http.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/clinic-vet-api_internal_infrastructure_http_handlers_dtos.SessionResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid credentials or validation error",
                        "schema": {
                            "$ref": "#/definitions/clinic-vet-api_internal_shared_http.APIResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/clinic-vet-api_internal_shared_http.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/clinic-vet-api_internal_shared_http.APIResponse"
                        }
                    }
                }
            }
        },
        "/auth/logout": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Invalidates the current session (refresh token). Requires Bearer auth.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Logout",
                "parameters": [
                    {
                        "description": "Refresh token to revoke",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/clinic-vet-api_internal_infrastructure_http_handlers_dtos.LogoutRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Logged out",
                        "schema": {
                            "$ref": "#/definitions/clinic-vet-api_internal_shared_http.APIResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/clinic-vet-api_internal_shared_http.APIResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/clinic-vet-api_internal_shared_http.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/clinic-vet-api_internal_shared_http.APIResponse"
                        }
                    }
                }
            }
        },
        "/auth/logout-all": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Invalidates all refresh tokens for the authenticated user. Requires Bearer auth.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Logout from all devices",
                "responses": {
                    "200": {
                        "description": "Logged out from all devices",
                        "schema": {
                            "$ref": "#/definitions/clinic-vet-api_internal_shared_http.APIResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/clinic-vet-api_internal_shared_http.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/clinic-vet-api_internal_shared_http.APIResponse"
                        }
                    }
                }
            }
        },
...
        "/auth/register": {
            "post": {
                "description": "Creates a new user account (admin, customer, or employee). Admins get immediate access; others may require activation.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Register a new user",
                "parameters": [
                    {
                        "description": "Registration data",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/clinic-vet-api_internal_infrastructure_http_handlers_dtos.RegisterRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Registration successful",
                        "schema": {
                            "$ref": "#/definitions/clinic-vet-api_internal_shared_http.APIResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request body or validation error",
                        "schema": {
                            "$ref": "#/definitions/clinic-vet-api_internal_shared_http.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/clinic-vet-api_internal_shared_http.APIResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "clinic-vet-api_internal_infrastructure_http_handlers_dtos.EnableTwoFactorRequest": {
            "description": "2FA method: totp, sms, or email.",
            "type": "object",
            "required": [
                "method"
            ],
            "properties": {
                "method": {
                    "type": "string",
                    "enum": [
                        "totp",
                        "sms",
                        "email"
                    ]
                }
            }
        },
...
        "clinic-vet-api_internal_shared_http.APIResponse": {
            "description": "Standardized API response structure.",
            "type": "object",
            "properties": {
                "data": {
                    "description": "The data payload of the response for successful requests."
                },
                "error": {
                    "description": "Details of the error if the request was not successful.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/clinic-vet-api_internal_shared_http.ErrorInfo"
                        }
                    ]
                },
                "message": {
                    "description": "A message providing additional context about the response.",
                    "type": "string"
                },
                "meta": {
                    "description": "Additional metadata for the response, such as pagination info."
                },
                "request_id": {
                    "description": "A unique identifier for the request.",
                    "type": "string"
                },
                "success": {
                    "description": "Indicates whether the request was successful.",
                    "type": "boolean"
                },
                "timestamp": {
                    "description": "The timestamp when the response was generated.",
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Bearer JWT token. Example: \"Bearer \u0026lt;access_token\u0026gt;\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}